Proposing change in title
Many customers use this Article to relay email from on-prem Application or servers to internet through Office 365. However the title of the Article is somewhat misleading. The change happened in early 2017. I propose a title that is more generic to this requirement and something that customers can find easily. Also removing the "Important Notice Part" which does not seem relevant anymore.