Update rebuild-an-office-web-apps-2013-farm-easily.md
Product names referenced in this article is incorrect. There is no such thing as Offices Online 2013. This documentation applies to Office Web Apps 2013, an On-Premise server product used to view and Edit Office files through the browser as documented here:

- Office Web Apps Server overview: https://docs.microsoft.com/en-us/webappsserver/office-web-apps-server-overview

The following articles need to be corrected to reflect the correct product names (Office Web Apps 2013 vs. Offices Online 2013):

- https://docs.microsoft.com/en-us/office/troubleshoot/administration/rebuild-an-office-web-apps-2013-farm-easily
- https://docs.microsoft.com/en-us/office/troubleshoot/administration/office-web-apps-and-vmware