Update cant-open-sp-documents-from-chrome.md
There is no fix for this. SharePoint's product group attempted to introduce a fix in August 2015, however this did not change the product behavior. SharePoint 2013 is in extended support and is no longer receiving feature changes or non-security related bug fixes. There is little to no chance that this will be fixed in the future for the classic SharePoint UI since development has moved onto the Modern SharePoint UI experiance.