msExchResourceCapacity and msExchResourceDisplay
Adding msExchResourceCapacity and msExchResourceDisplay attribute value causes warning to be displayed while attempting to view the properties of such resource mailbox in OnPremise Servers.

Error message reported is:
WARNING: The object contoso.lab/Hybrid Mailboxes/testroom has been corrupted or isn't compatible with Microsoft
support requirements, and it's in an inconsistent state. The following validation errors happened:
WARNING: The type of the resource mailbox can't be blank. It should be either Room or Equipment.

From my understanding this error happens due to validation of object when we attempt to display the user property and it shows failure. Removing these properties fixes the error.

Would be better to have documented that they should not be managing the Room or Equipment mailboxes from OnPremise instead should be doing it from Exchange Online as a disclaimer. 

Note: You may run into warning while you access such room or equipment mailboxes if they are attempted to be accessed using EAC or Exchange Management Shell in OnPremise server. In case this needs to be modified some properties at the OnPremise server, clear these attributes and perform the intended task using Exchange Management Shell or EAC. Upon completion you can update `msExchResourceCapacity` and  `msExchResourceDisplay` back to original values.