Update rest-api-is-not-yet-supported-for-this-mailbox-error.md
These changes had been made to the article version in Portico in Sept 2020 but the migration was done using a previous version of the article that had been converted to .md using the conversion tool. The verbiage in this article was wrong and caused problems in Sept and the same problem has re-surfaced now because of the inaccuracy in the content. Hence fixing this version.