Update default-folder-is-missing.md
Added a note above the symptoms header to indicate that this article does not apply to REST-enabled Shared Calendars because they have the PR_ATTR_HIDDEN MAPI property set to true by design and that design does not cause problems with accessing the Shared Calendar as this article suggests. This article is still applicable to MAPI-based folders.