Update o365-group-tasks.md
S500 customer (MASTERCARD) on case 2408140040010765001 has requested to update this documentation as the operation can take up to 48 hours to be completed. 

The customer would like to have the documentation explicitly saying 48 hrs, or up to 2 days:

Customer statement:
Currently, the document states "several hours" instead of "up to 48hrs" or "up to 2 days".